requirements.txt
Flask==3.1.0
Appium-Python-Client==4.5.1
selenium==4.29.0
robotframework==7.2.2
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0